fix formats that explicitly test for particular characters
base32, base64, and ascii85 all need to check for particular characters:
CONVERTER logic in the decoding functions was misinterpreted for this
case.  CONVERTER is really for turning elements of the input sequence
into integers and was inappropriately being used to try to convert the
special characters into integers, which didn't work when decoding octet
vectors.

For those checks, explicitly check the type of sequence we're decoding
from and take the appropriate transformation on the special character.